Update README.md
Shell money is a form of currency that was used in various parts of the world, including Asia, Africa, and Oceania. This type of money makes use of a type of marine snail known as cowrie and therefore is known also as cowrie shell money. In some parts of the world shell money served as currency up until the 19th/20th century.